requirements.txt
# Agent framework and Semantic search engine
Agno>=2.2.12
qdrant-client>=1.15.1

# LLMs
anthropic>=0.72.1
google-genai>=1.50.1
openai>=2.8.0

# Image handling
pdf2image>=1.17.0
pillow>=12.0.0

# Document processing
chonkie[all]>=1.4.1

#embedding
fastembed>=0.7.3

# LlamaIndex for vector storage and indexing
llama-index-core>=0.14.7
llama-index-llms-openai>=0.6.7
llama-index-embeddings-openai>=0.5.1
llama-index-embeddings-google-genai>=0.3.1
llama-index-embeddings-ollama>=0.8.3
llama-index-vector-stores-qdrant>=0.8.6
llama-index-embeddings-fastembed>=0.5.0

# misc
python-dotenv>=1.2.1
googlesearch-python>=1.3.0
pycountry>=24.6.1
